Move billing details into its own component
